P2866 P6510

+++ P2866.cpp
@@ -9,14 +9,14 @@ int main() {
     int n; std::cin >> n;
     int *inp = (int*) malloc(sizeof(int) * n);
     for (int i = 0; i < n; i++) std::cin >> inp[i];
-    int ans = 0;
+    long long ans = 0;
     std::stack<int> stack;
     for (int i = 0; i < n; i++) {
         while ((!stack.empty()) && (stack.top() <= inp[i])) stack.pop();
         ans += stack.size();
         stack.push(inp[i]);
     }
-    printf("%d\n", ans);
+    printf("%lld\n", ans);
     free(inp);
     return 0;
 }
